hive/health: move healthv2 to hive and rename it
Moves the healthv2 implementation to hive/health to get rid of the
somewhat ugly v2 part of the pkg name, and to give some logical
structure.
