--- pkg/healthv2/types.go
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-package healthv2
